✨ feat: implement course existence check to verify course existence prior to deletion
- Add checkCourses function in coursesDataAccess.ts to verify course existence before deletion occurs.
- Update CoursesService to incorporate this check and ensure that deletion requests for non-existent courses are not carried out.
- Modify deleteCourses function in coursesController.ts to return a 404 error when a deletion is attempted on a non-existent course, informing users about the course's non-existence.

+++ src/service/CoursesService.ts
@@ -18,6 +18,10 @@ export class CoursesService {
   }
 
   async deleteCourses(req: Request) {
+    const courseExists: number = await dataAccess.checkCourses(this.pool, req);
+    if (!courseExists) {
+      throw new Error("Kursen med det Id't existerar inte!");
+    }
     return dataAccess.deleteCourses(this.pool, req);
   }
 }
